Internal: use dynamic imports to load players
