+++ packages/studio-base/src/players/buildPlayer.ts
@@ -5,7 +5,12 @@
 import RandomAccessPlayer from "@foxglove/studio-base/players/RandomAccessPlayer";
 import { Player, PlayerMetricsCollectorInterface } from "@foxglove/studio-base/players/types";
 import { CoreDataProviders } from "@foxglove/studio-base/randomAccessDataProviders/constants";
+import {
+  getLocalBagDescriptor,
+  getRemoteBagDescriptor,
+} from "@foxglove/studio-base/randomAccessDataProviders/standardDataProviderDescriptors";
 import { RandomAccessDataProviderDescriptor } from "@foxglove/studio-base/randomAccessDataProviders/types";
+import { SECOND_SOURCE_PREFIX } from "@foxglove/studio-base/util/globalConstants";
 import { getSeekToTime } from "@foxglove/studio-base/util/time";
 
 export type BuildPlayerOptions = {
@@ -35,4 +40,59 @@ export function buildPlayerFromDescriptor(
     metricsCollector: options.metricsCollector,
     seekToTime: getSeekToTime(),
   });
+}
+
+export function buildPlayerFromFiles(files: File[], options: BuildPlayerOptions): Player {
+  const name = files.map((file) => String(file.name)).join(", ");
+  if (files.length === 1) {
+    return buildPlayerFromDescriptor(name, getLocalBagDescriptor(files[0] as File), options);
+  } else if (files.length === 2) {
+    return buildPlayerFromDescriptor(
+      name,
+      {
+        name: CoreDataProviders.CombinedDataProvider,
+        args: {},
+        children: [
+          getLocalBagDescriptor(files[0] as File),
+          {
+            name: CoreDataProviders.RenameDataProvider,
+            args: { prefix: SECOND_SOURCE_PREFIX },
+            children: [getLocalBagDescriptor(files[1] as File)],
+          },
+        ],
+      },
+      options,
+    );
+  }
+  throw new Error(`Unsupported number of files: ${files.length}`);
+}
+
+export function buildPlayerFromBagURLs(urls: string[], options: BuildPlayerOptions): Player {
+  const name = urls.map((url) => url.toString()).join(", ");
+
+  if (urls.length === 1) {
+    return buildPlayerFromDescriptor(
+      name,
+      getRemoteBagDescriptor(urls[0] as string, options),
+      options,
+    );
+  } else if (urls.length === 2) {
+    return buildPlayerFromDescriptor(
+      name,
+      {
+        name: CoreDataProviders.CombinedDataProvider,
+        args: {},
+        children: [
+          getRemoteBagDescriptor(urls[0] as string, options),
+          {
+            name: CoreDataProviders.RenameDataProvider,
+            args: { prefix: SECOND_SOURCE_PREFIX },
+            children: [getRemoteBagDescriptor(urls[1] as string, options)],
+          },
+        ],
+      },
+      options,
+    );
+  }
+  throw new Error(`Unsupported number of urls: ${urls.length}`);
 }
